:sparkles: [BE] 아이템 사용 로그 저장, 버그 수정

+++ api-server/LogServer/src/main/java/com/exodia/logserver/db/repository/InGameClearLogRepository.java
@@ -1,8 +1,10 @@
 package com.exodia.logserver.db.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.exodia.logserver.db.entity.InGameClearLog;
 
-public interface InGameClearLogRepository extends MongoRepository<InGameClearLog, Long> {
+public interface InGameClearLogRepository extends MongoRepository<InGameClearLog, BigInteger> {
 }
